Refactored Filter stories organization.
