+++ src/stories/Filter/Filter.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
 
-import FilterPage from './FilterPage';
+import FilterPage from './components/FilterPage';
 import FilterComponent from 'src/Filter/Filter';
 import DocBlock from '.storybook/DocBlock';
