+++ src/stories/Filter/components/index.ts
@@ -1,5 +1,6 @@
 export { default as ContextSwitchMenu } from './ContextSwitchMenu';
 export { default as FilterBar } from './FilterBar';
+export { default as FilterPage } from './FilterPage';
 export { default as FilterJSONConfirmationModal } from './FilterJSONConfirmationModal';
 export { default as FilterJSONInput } from './FilterJSONInput';
 export { default as UrlPreview } from './UrlPreview';
